+++ R/catalog_index.r
@@ -33,8 +33,9 @@ catalog_index =	function(catalog, bboxes, shape = LIDRRECTANGLE, buffer = 0)
 
   queries <- lapply(bboxes, function(bbox)
   {
-    bbox = bbox + 2*buffer
-    files = catalog@data[!( `Min X` >= bbox@xmax | `Max X` <= bbox@xmin | `Min Y` >= bbox@ymax | `Max Y` <= bbox@ymin)]$filename
+    bbox  = bbox + 2*buffer
+    is_in = with(catalog@data, !( `Min X` >= bbox@xmax | `Max X` <= bbox@xmin | `Min Y` >= bbox@ymax | `Max Y` <= bbox@ymin))
+    files = catalog@data$filename[is_in]
 
     if (length(files) == 0)
       return(NULL)