+++ tests/testthat/test-catalog_retile.R
@@ -1,19 +1,14 @@
 context("catalog_retile")
 
 lidr_options(interactive = FALSE)
-folder <- system.file("extdata", "", package="lidR")
+folder <- system.file("extdata", "Megaplot.laz", package="lidR")
 ctg = catalog(folder)
 cores(ctg) <- 1
+buffer(ctg) <- 0
 tiling_size(ctg) = 80
 progress(ctg) <- FALSE
 
 test_that("catalog reshape works", {
-  ctg = catalog(folder)
-  tiling_size(ctg) = 80
-  buffer(ctg) = 0
-  cores(ctg) = 1
-  progress(ctg) <- FALSE
-  ctg@data = ctg@data[1]
   temp = tempfile()
 
   ctg2 = catalog_retile(ctg, temp, prefix = "test_")